requirements.txt
2captcha-python==1.5.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
Appium-Python-Client==4.4.0
attrs==24.2.0
beautifulsoup4==4.13.4
blinker==1.9.0
Brotli==1.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
Deprecated==1.2.18
Flask==3.1.0
Flask-Limiter==3.12
frozenlist==1.7.0
funkify==0.4.5
h11==0.14.0
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kaitaistruct==0.10
limits==5.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
MouseInfo==0.1.3
msgspec==0.19.0
multidict==6.6.3
numpy==2.1.3
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
propcache==0.3.2
pyasn1==0.6.1
PyAutoGUI==0.9.54
pycparser==2.22
pydivert==2.1.0
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==1.0.9
pyOpenSSL==24.3.0
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-2captcha==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.2
requests==2.32.3
requires==0.10.5
retrying==1.3.4
rich==13.9.4
selenium==4.34.2
selenium-wire==5.1.0
setuptools==80.9.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
tenacity==9.1.2
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2024.2
urllib3==2.5.0
urllib3-secure-extra==0.1.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.3
wsproto==1.2.0
xtyping==0.8.2
yarl==1.20.1
zstandard==0.23.0
